save midi file functionality

+++ backend/app/schemas/project.py
@@ -16,8 +16,8 @@ class TrackBase(BaseModel):
 class Track(TrackBase):
     id: str
     duration: Optional[float] = None
-    x_position: int = 0
-    y_position: int = 0
+    x_position: float = 0.0  # Changed to float to support partial positions
+    y_position: float = 0.0  # Changed to float for consistency
     storage_key: str
     left_trim_ms: int = 0
     track_number: int = 0
